Bug 1722127 - Modify rlbox graphite struct defs for windows compat r=tjr
Differential Revision: https://phabricator.services.mozilla.com/D120777

UltraBlame original commit: eae907d9030ca5d65bb021265ec031a3f374334a

+++ gfx/graphite2/geckoextra/include/GraphiteStructsForRLBox.h
@@ -19,14 +19,17 @@
 
 #endif
 
+
+
+
 #define sandbox_fields_reflection_graphite_class_gr_font_ops(f, g, ...)                       \
-  f(size_t, size, FIELD_NORMAL, ##__VA_ARGS__) g()                                            \
+  f(unsigned int, size, FIELD_NORMAL, ##__VA_ARGS__) g()                                            \
   f(float (*)(const void*, unsigned short), glyph_advance_x, FIELD_NORMAL, ##__VA_ARGS__) g() \
   f(float (*)(const void*, unsigned short), glyph_advance_y, FIELD_NORMAL, ##__VA_ARGS__) g()
 
 #define sandbox_fields_reflection_graphite_class_gr_face_ops(f, g, ...)                              \
-  f(size_t, size, FIELD_NORMAL, ##__VA_ARGS__) g()                                                   \
-  f(const void* (*)(const void*, unsigned int, size_t*), get_table, FIELD_NORMAL, ##__VA_ARGS__) g() \
+  f(unsigned int, size, FIELD_NORMAL, ##__VA_ARGS__) g()                                                   \
+  f(const void* (*)(const void*, unsigned int, unsigned int*), get_table, FIELD_NORMAL, ##__VA_ARGS__) g() \
   f(void (*)(const void*, const void*), release_table, FIELD_NORMAL, ##__VA_ARGS__) g()
 
 #define sandbox_fields_reflection_graphite_class_gr_glyph_to_char_cluster(f, g, ...) \
